+++ lib/ncs_navigator/warehouse/transformers.rb
@@ -11,17 +11,11 @@ module NcsNavigator::Warehouse
   # `Enumerable` of MDES Warehouse model instances, validates and
   # saves them. It's a good general base for your own transformers.
   module Transformers
-    autoload :ApplyGlobalValuesFilter, 'ncs_navigator/warehouse/transformers/apply_global_values_filter'
-    autoload :CodedAsMissingFilter,    'ncs_navigator/warehouse/transformers/coded_as_missing_filter'
     autoload :Database,                'ncs_navigator/warehouse/transformers/database'
     autoload :EnumTransformer,         'ncs_navigator/warehouse/transformers/enum_transformer'
     autoload :EventStartFromContactTransformer, 'ncs_navigator/warehouse/transformers/event_start_from_contact_transformer'
-    autoload :Filters,                 'ncs_navigator/warehouse/transformers/filters'
     autoload :ForeignKeyIndex,         'ncs_navigator/warehouse/transformers/foreign_key_index'
     autoload :MdesCsv,                 'ncs_navigator/warehouse/transformers/mdes_csv'
-    autoload :NoBlankForeignKeysFilter,       'ncs_navigator/warehouse/transformers/no_blank_foreign_keys_filter'
-    autoload :NoSsuOutreachAllSsusFilter,     'ncs_navigator/warehouse/transformers/no_ssu_outreach_all_ssus_filter'
-    autoload :NoSsuOutreachPlaceholderFilter, 'ncs_navigator/warehouse/transformers/no_ssu_outreach_placeholder_filter'
     autoload :SamplingUnits,           'ncs_navigator/warehouse/transformers/sampling_units'
     autoload :SqlTransformer,          'ncs_navigator/warehouse/transformers/sql_transformer'
     autoload :SubprocessTransformer,   'ncs_navigator/warehouse/transformers/subprocess_transformer'