+++ spec/ncs_navigator/warehouse/filters/no_blank_foreign_keys_filter_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-module NcsNavigator::Warehouse::Transformers
+module NcsNavigator::Warehouse::Filters
   describe NoBlankForeignKeysFilter, :use_mdes do
     describe '.call' do
       def call(records)